+++ index.ts
@@ -12,4 +12,4 @@ export { stubsRoot } from './stubs/main.ts'
 
 export { ModalResponse } from './src/modal_response.ts'
 export { ModalHeaders } from './src/headers.ts'
-export type { ModalPayload, ModalProps } from './src/types.ts'
+export type { Backdrop, ModalPayload, ModalProps } from './src/types.ts'